+++ matrix.c
@@ -6,7 +6,7 @@
 /*   By: fmaurer <fmaurer42@posteo.de>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/10 21:33:41 by fmaurer           #+#    #+#             */
-/*   Updated: 2024/09/12 07:08:41 by fmaurer          ###   ########.fr       */
+/*   Updated: 2024/09/12 11:22:30 by fmaurer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,9 @@
 /* mat * vec. works in-place. */
 void	mult_mat_vec(double a[3][3], t_vec *v)
 {
-	v->x = (int)((double)v->x * a[0][0] + (double)v->y * a[0][1] + (double)v->z * a[0][2]);
-	v->y = (int)((double)v->x * a[1][0] + (double)v->y * a[1][1] + (double)v->z * a[1][2]);
-	v->z = (int)((double)v->x * a[2][0] + (double)v->y * a[2][1] + (double)v->z * a[2][2]);
+	v->x = v->x * a[0][0] + v->y * a[0][1] + v->z * a[0][2];
+	v->y = v->x * a[1][0] + v->y * a[1][1] + v->z * a[1][2];
+	v->z = v->x * a[2][0] + v->y * a[2][1] + v->z * a[2][2];
 }
 
 void	mult_mat_map(double a[3][3], t_map *map)
